Resolve some redundancies on codes, type some methods, fix code format.
Eliminate double negation from JupyterServersList.loading.

Eliminate double negation from JupyterServersList.isTarget.

Supply open and close curly braces lacked in a comment.

Remove a redundant slash in a closing tag.

Rename translation properties. The entry is used if one or more of
following files are dirty.
  + Dockerfile
  + environment.yml
  + requirements.txt
Namely, it is not limited to Dockerfile.

Include BinderHubConfigModel in the model of guid-node/binderhub route.