Show loading message on ProjectEditor a bit longer.
The `loading` property became `true` too early. The template first
refers `manuallyChanged` and iff that property is `false`,
`selectedImageUrl` property is referred.  The asynchronously process
`loadCurrentConfig` and `mergeConfigurations`, called in
`didReceiveAttrs` must finish beforehand in order to get the correct
value of this property. However, this `loading` property turned into
`false` if it could see `dockerfile` property, which is set as a *part*
of `loadCurrentConfig`. In other words, the existance of `dockerfile`
property does NOT denote the initialization process finishes.

This commit fixes this situation by introducing a new boolean property,
`isInitialized`, and set the property after `loadCurrentConfig` and
`mergeConfigurations` have finished.