Introduce findConfigurationFile method to pick a ConfigurationFile object.
Not only the implementation, but also the existance of this method
itself should be considered as an intermediate one. The
ConfigurationFile should have an `id` property so we can search safely
find over the `configurationFiles` property. The name
`ConfigurationFile` is not proper since it is not a file but something
that includes versatile information of a configuration file (it even
includes a property name of the dirty bit!). Even more, that property
should be an object instead of an array since we need to pick one actual
ConfigurationFile object in several cases.  The use of this method make
it easier to fix (or, resolve) these concerns.