Save Dockerfile content before launch.
This implementaion is not satisfying. It gains huge amount of code
complexity to insert a small process before launch. The best way I can
imagine is to click the save button on the Dockerfile Editor. Since
BuildConsole and ProjectEditor (which holds the Dockerfile Editor) are
siblings, we need some tricks. I expect that we can implement a tiny
modifier to do this. This approach access to DOM; call the `click`
method of the `input` element. In order to access DOM in Ember.js, we
should use modifier.

Another, standard way is to lifting up the states to their common
ancestor, namely, the controller of the page. In this approach, we need
to move so many lines of ProjectEditor to the controller.