Introduce a controller property to hold server annotations state.
Finally, Route models are only useful to intialize the state variables,
not to hold the state. There is less chance to update it without *whole*
page re-rendering since there is no clue to update a *part* of a it. I
therefore decided to use it only for initialization of Controller's
property, `dyServerAnnotations` in concrete. The same approach should
work fine for `binderHubConfig` too, and I guess that, the combination
of `config` getter and `configCache` might be introduced to achieve the
same thing. But the code is collection of ad-hocs and quite hard to
understand the intention. It should be refactored.

About `node` property, the situation _seems_ more complexed. I cannot
understande why it is written with using ember-concurrency. I am not
sure if it can be refactored by the same approach.

Partial update of models might be possible if we (fully) migrate to
Octane and utilize `@tracked`.