Add MATLAB release picker UI and product picker UI.
Replace MPM release editor by a list UI.

Add Add-ons (i.e. products) picking modal window.

Add products filter on MATLAB products picking modal dialogue.

Highlight the text portion which reasons why the name passes the filter.

Avoid to break the horizontal boundary by the list of picked products.

Adopt multiple text fragments to filter MATLAB producsts.

Make MATLAB product name filtering case insensitive.

Add a feature that allows to specify MATLAB products name manually.

Prevent to add one MATLAB product 2 or more times.

Disable the ADD button on the manually add tab if the name is empty.

Prevent to add MATLAB package whose name is an empty string.

Suppress the malformed behaviour on mouse hover over the tab.
.nav-tabs > li > a:hover is set `border-bottom: none`, while the
thickness of border is 1px if it is not hovered. As a result, the height
of tab changed by 1px if it is mouse-overed. In our MATLAB products
picking dialogue case, if the last tab of the first tab line, namely the
`O` tab, is hovered, then suddenly a new tab line is inserted between
existing 2 tab lines and `P` and `R` tabs move to that new line. This
behaviour is weird and need to be resolved. Even though this deeply
nested CSS selector violates the stylelinter, this kind of workaround
should be carried out in narrower range as much as possible.

Acquire MATLAB product names from the server, in on-the-fly manner.

Prevent to flush the picked products on page initialization.

Reject manual addon name if it includes whitespaces in the middle.

Fix and add some test cases.