Restrict the binderhub feature if BinderHub/JupyterHub is not available.
Disable the Jupyter Servers list and the build button if BinderHub
and/or JupyterHub are not responding or returns errorly status codes.
By preventing to use such unavailable BinderHub/JupyterHub servers,
the binderhub page itself appears without the browser's error page.
The parts of the page other than jupyter servers list and the build
button are left normal.

Warning banner is shown on top of the binderhub page if the health check
endpoints of BinderHub and/or JupyterHub are not available. Status code
is shown on Toast if Hubs are not available.

Following changes are also included:

+ Add some test cases
+ Change the heading label on the host info component